universial.py renaming fixes
