+++ sparrowpy/form_factor/form_factor.py
@@ -7,7 +7,7 @@ except ImportError:
     prange = range
 import numpy as np
 from sparrowpy.radiosity_fast import geometry
-from sparrowpy.form_factor.universal_methods import ( calc_form_factor )
+from sparrowpy.form_factor.universial import ( calc_form_factor )
 
 
 def kang(
